+++ Engine/Source/Raylib/RaylibInputWrapper.h
@@ -1,6 +1,6 @@
-#include <array>
-
 #pragma once
+
+#include <array>
 
 bool IsKeyPressedWrapper(int key);
 bool IsKeyReleasedWrapper(int key);
@@ -10,8 +10,8 @@ bool IsMouseButtonPressedWrapper(int button);
 bool IsMouseButtonReleasedWrapper(int button);
 bool IsMouseButtonDownWrapper(int button);
 void SetMousePositionWrapper(int x, int y);
-std::array<int, 2> GetMousePositionWrapper();
-std::array<int, 2> GetMouseDeltaWrapper();
+std::array<float, 2> GetMousePositionWrapper();
+std::array<float, 2> GetMouseDeltaWrapper();
 void SetMouseScaleWrapper(float x, float y);
 void SetMouseCursorWrapper(int cursor);
 float GetMouseWheelMoveWrapper();
